+++ CHANGELOG.md
@@ -17,6 +17,9 @@ T.t = 1
 local a = t -- hover on t show two same t definition
 ```
 
+2. fix `...` param grammars injection invalid
+
+
 ## [1.0.13] 2020-08-17
 1. fix grammars injection